Test I2 C line
The I2C line was tested with the new PCB and it was found that the I2C
bus would not return data.  Test data is added and analysed.

TO DO
- amend clock rate of the I2C bus.  There appears to be extra
capacitance in the line.  Attempt to amend this by reducing clock rate
of the I2C bus.  At present the arduino is using the default rate of
100 kHz.  The ATmega 328 IC supports data rates of up to 400 KHz with
no set minimum.  The pull-up resistors within the arduino are 20k.  The
clock rate of the I2C bus can be set within the Wire library by by
calling Wire.setclock(X);  in the range 31000L < X< 400000L where X is
the bitrate in Hz e.g. Wire.setclock(400000L); sets the bitrate to 400
kHz.  This should be called after the Wire.begin(); command as this
initialises the bitrate to 100 kHz.